Move prime fields into prime/
