+++ libff/algebra/fields/prime/fp.tcc
@@ -14,8 +14,8 @@
 #include <cstdlib>
 #include <limits>
 
-#include <libff/algebra/fields/field_utils.hpp>
-#include <libff/algebra/fields/fp_aux.tcc>
+#include <libff/algebra/fields/prime/field_utils.hpp>
+#include <libff/algebra/fields/prime/fp_aux.tcc>
 
 namespace libff {
