+++ libff/algebra/fields/prime/fp4.hpp
@@ -16,8 +16,8 @@
 #ifndef FP4_HPP_
 #define FP4_HPP_
 
-#include <libff/algebra/fields/fp.hpp>
-#include <libff/algebra/fields/fp2.hpp>
+#include <libff/algebra/fields/prime/fp.hpp>
+#include <libff/algebra/fields/prime/fp2.hpp>
 
 namespace libff {
 
@@ -99,6 +99,6 @@ Fp_model<n, modulus> Fp4_model<n, modulus>::Frobenius_coeffs_c1[4];
 
 } // libff
 
-#include <libff/algebra/fields/fp4.tcc>
+#include <libff/algebra/fields/prime/fp4.tcc>
 
 #endif // FP4_HPP_
